move modal into own file
